rtw-move-tokenize-helpers: Hand tokenize-pipeline helpers over to RawTokenizerWrapper
Pure cut/paste move per MECH_COMMIT_SPLIT. Cuts 4 @staticmethod helpers
(``_detect_input_format`` / ``_prepare_tokenizer_input`` /
``_extract_tokenizer_results`` / ``_tokenize_texts``) from TM; pastes
into RawTokenizerWrapper (drop @staticmethod, restore plain ``self``,
collapse intra-cluster ``TokenizerManager._x(self, ...)`` calls back to
``self._x(...)`` — they are sibling instance methods on the new owner).

External caller prefix replacement in TM:
``TokenizerManager._helper(self.raw_tokenizer_wrapper, ...)`` →
``self.raw_tokenizer_wrapper._helper(...)``.

Refactor chain ID: rtw-move-tokenize-helpers
